Fixed type sizes in ILI9341 library

+++ screen/Inc/ili9341_gfx.h
@@ -117,7 +117,7 @@ void ili9341_spi_tft_set_address_rect(ili9341_t *lcd,
     uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1);
 
 void ili9341_transmit_wait(ili9341_t *lcd);
-void ili9341_transmit_color(ili9341_t *lcd, uint16_t size,
+void ili9341_transmit_color(ili9341_t *lcd, uint32_t size,
     uint16_t color[]/* already byte-swapped (LE) */, ili9341_bool_t wait);
 
 void ili9341_draw_pixel(ili9341_t *lcd, ili9341_color_t color,
